Move `launch_beam` into new file
